# api/openapi.yaml
openapi: 3.0.3
info:
  title: nfrx API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
paths:
...
  /api/v1/chat/completions:
    post:
      security: [ { BearerAuth: [] } ]
      summary: Chat completions (proxy)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
            text/event-stream:
              schema:
                type: string
...